Bug fixes to site navigation
Removed the “/“ between the links on navigaton from inline code. The slash now appears thanks to an inclusion of the css content property.

Additionally, because only two links need the navigation for the desktop view (and this looks hella weird on tablet/mobile views) added this to a media query.